+++ src/components/TimerContainer/TimerContainer.tsx
@@ -1,10 +1,7 @@
-import { useState } from 'react';
 import Image from 'next/image';
 import { css, Theme } from '@emotion/react';
 
-import { Button } from '~/components/Button';
-import CloseIcon from '~/components/Icons/CloseIcon';
-import Modal from '~/components/Modal/Modal';
+import GuideModalButton from '~/components/GuideModalButton';
 // import { END_DATE } from '~/constant/common';
 // import useIsInProgress from '~/hooks/useIsInProgress';
 import { commonLayoutCss } from '~/styles/layout';
@@ -21,14 +18,6 @@ export function TimerContainer() {
   // const onButtonClick = () => {
   //   router.push('/apply');
   // };
-
-  const [isAlertOpen, setIsAlertOpen] = useState(false);
-
-  const onAlertClose = () => setIsAlertOpen(false);
-
-  const onAlertButtonClick = () => {
-    setIsAlertOpen(true);
-  };
 
   return (
     <section css={containerCss}>
@@ -49,29 +38,12 @@ export function TimerContainer() {
               <p>완성하며 성장하는 IT 커뮤니티입니다</p>
             </div>
           </div>
-          <Button size="lg" overrideCss={buttonCss} onClick={onAlertButtonClick}>
-            15기 알림신청
-          </Button>
           {/* <Timer time={progressState === 'FINISH' ? FINISH_TIME_OBJ : time} />
           <Button size="lg" disabled={!isInProgress} onClick={onButtonClick}>
             14기 지원하기
           </Button> */}
 
-          <Modal isShowing={isAlertOpen} onClickOutside={onAlertClose} mode="wait">
-            <article css={modalCss}>
-              <h2>15기 알림 신청</h2>
-              <p>디프만 카카오톡 채널 친구 추가 시 기수 모집 알림을 보내드립니다</p>
-              <Image src={'/images/kakao-qr.png'} alt="kakao qr link" width={90} height={90} />
-              <div css={buttonWrapperCss}>
-                <a href="https://pf.kakao.com/_xoxmcxed">
-                  <Button size="md">카카오톡 채널 바로가기</Button>
-                </a>
-              </div>
-              <div className="close-icon" onClick={onAlertClose}>
-                <CloseIcon />
-              </div>
-            </article>
-          </Modal>
+          <GuideModalButton />
         </div>
       </div>
     </section>
@@ -84,63 +56,6 @@ export function TimerContainer() {
 //   min: '00',
 //   sec: '00',
 // };
-
-const buttonWrapperCss = css`
-  margin-top: 39px;
-  margin-bottom: 24px;
-  padding: 0 24px;
-  width: 100%;
-
-  button {
-    height: 46px;
-    width: 100%;
-
-    ${mediaQuery('mobile')} {
-      height: 46px;
-    }
-  }
-`;
-
-const modalCss = (theme: Theme) => css`
-  background-color: ${theme.colors.black400};
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  width: 320px;
-  position: relative;
-
-  h2 {
-    ${theme.typos.pretendard.body1};
-    color: ${theme.colors.white};
-    margin-bottom: 8px;
-    margin-top: 47px;
-  }
-
-  p {
-    ${theme.typos.pretendard.body2};
-    color: ${theme.colors.gray100};
-    margin-bottom: 39px;
-    max-width: 221px;
-  }
-
-  .close-icon {
-    position: absolute;
-    top: 16px;
-    right: 16px;
-    cursor: pointer;
-  }
-`;
-
-const buttonCss = css`
-  max-width: 443px;
-  margin: 0 auto;
-
-  ${mediaQuery('mobile')} {
-    max-width: 266px;
-  }
-`;
 
 const containerCss = css`
   position: relative;
